feat(subscriptions): Make --override-result-topic optional in executor (#2731)
This was being used during testing to ensure the executor wasn't producing
results to the real topic. Now we are going to use the real topic from the
stream loader.

This is the first step towards removing it completely. Once we are no
longer passing this anywhere, it will be removed completely.