feat(subscriptions): Improve filtering of stale subscriptions in scheduler (#2765)
During a multi-hour outage, the scheduler gets into the unusual state of having
ticks spanning multiple hours, for which a very large number of subscriptions
need to be scheduled (often exceeding the buffer size of 10,000 applied later
in the pipeline). This is likely to be a cause of the scheduler getting stuck.

If --stale-threshold-seconds is passed, we should use the lower bound of the tick
instead of the upper bound for comparison so that extremely large ticks that are
not able to be properly processed get discarded.